+++ FootPrints/src/main/java/footprints/footprints/domain/post/PostDTO.java
@@ -23,8 +23,8 @@ public class PostDTO {
     private String valid_time;       // 게시물 유효 시간
     private int view_num;         // 조회수
     private String nick;
-    private long x;
-    private long y;
+    private double x;
+    private double y;
 
 
     private Member member;  //0725추가
